misc style and type fix.

+++ compat.c
@@ -1427,7 +1427,7 @@ printf("Enter state PARSE_CONV_SPEC\n");
 						SET_FLAG(FLAG_ALTERNATIVE_FORM);
 */
 
-						convert_int(&data, (unsigned_t) (unsigned int) ptr);
+						convert_int(&data, (unsigned_t) (size_t) ptr);
 						break;
 
 					case 'n' :
